Add option to convert position into Area ID (#197)
This adds another command line flag to automatically convert the given
position into an Area ID (`--pos-to-area-id`). If either this new flag
or the `--area-id` argument are given then the CRA message with Area ID
is sent. I also changed several of the uses of `f64` to use `f32`
instead to make sure there weren't rounding issues in the Area ID
calculations.

Is this an acceptable way to squeeze in this functionality? I was also
thinking about adding these functions into
https://github.com/swift-nav/swiftnav-rs, but that's not a dependency
for ntripping and would potentially complicate the portability since it
pulls in C code as well.

---------

Co-authored-by: Patrick Crumley <patrick.crumley@gmail.com>