Update Dependabot to ignore patch updates (#271)
We tend to get a bunch of patch updates to dependencies , which can
create a lot of pointless noise in the commit history. Since we have
security scans enabled, it should be fine to skip over patch updates
until we see encounter a bug that requires an update.

This also adds a reviewer to the Dependabot PRs, so that we can get them
reviewed and merged in a timely fashion.

---------

Co-authored-by: Patrick Crumley <patrick.crumley@swift-nav.com>